Why not the vdom provide communicating constructs
